requirements.txt
python-escpos==3.0a8
twitchio==1.1.0